# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "DataFreshnessResponse": {
        "properties": {
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Freshness"
            },
            "title": "Sources",
            "type": "array"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "sources"
        ],
        "title": "DataFreshnessResponse",
        "type": "object"
      },
...
      "FilingDetail": {
        "properties": {
          "accession_number": {
            "title": "Accession Number",
            "type": "string"
          },
          "company_ticker": {
            "title": "Company Ticker",
            "type": "string"
          },
          "filing_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filing Date"
          },
          "filing_id": {
            "title": "Filing Id",
            "type": "string"
          },
          "filing_type": {
            "title": "Filing Type",
            "type": "string"
          },
          "items": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Items"
          },
          "period_end": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period End"
          },
          "primary_document_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Document Url"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          }
        },
        "required": [
          "filing_id",
          "filing_type",
          "accession_number",
          "company_ticker"
        ],
        "title": "FilingDetail",
        "type": "object"
      },
      "FilingSection": {
        "properties": {
          "char_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Char End"
          },
          "char_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Char Start"
          },
          "section_path": {
            "items": {
              "type": "string"
            },
            "title": "Section Path",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "section_path",
          "title"
        ],
        "title": "FilingSection",
        "type": "object"
      },
      "FilingSectionsResponse": {
        "properties": {
          "filing_id": {
            "title": "Filing Id",
            "type": "string"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/FilingSection"
            },
            "title": "Sections",
            "type": "array"
          }
        },
        "required": [
          "filing_id",
          "sections"
        ],
        "title": "FilingSectionsResponse",
        "type": "object"
      },
...
      "Freshness": {
        "description": "Data-freshness metadata attached to data endpoints (API-008, FR-006).",
        "properties": {
          "as_of": {
            "description": "ISO-8601 UTC timestamp of the underlying data",
            "title": "As Of",
            "type": "string"
          },
          "freshness_status": {
            "enum": [
              "fresh",
              "stale",
              "failed"
            ],
            "title": "Freshness Status",
            "type": "string"
          },
          "source": {
            "description": "Provider or system that produced the data",
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "as_of",
          "source",
          "freshness_status"
        ],
        "title": "Freshness",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Contract-first API for InvestiLens (spec \u00a724).",
    "title": "InvestiLens API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/filings/{filing_id}": {
      "get": {
        "operationId": "get_filing_api_v1_filings__filing_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "filing_id",
            "required": true,
            "schema": {
              "title": "Filing Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilingDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Filing",
        "tags": [
          "Filings"
        ]
      }
    },
    "/api/v1/filings/{filing_id}/diff": {
      "get": {
        "description": "Filing-to-filing diff (Phase 6b). Not yet available.",
        "operationId": "diff_filings_api_v1_filings__filing_id__diff_get",
        "parameters": [
          {
            "in": "path",
            "name": "filing_id",
            "required": true,
            "schema": {
              "title": "Filing Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "against",
            "required": true,
            "schema": {
              "title": "Against",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Diff Filings",
        "tags": [
          "Filings"
        ]
      }
    },
    "/api/v1/filings/{filing_id}/sections": {
      "get": {
        "operationId": "get_filing_sections_api_v1_filings__filing_id__sections_get",
        "parameters": [
          {
            "in": "path",
            "name": "filing_id",
            "required": true,
            "schema": {
              "title": "Filing Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilingSectionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Filing Sections",
        "tags": [
          "Filings"
        ]
      }
    },
...
    "/api/v1/meta/data-freshness/{ticker}": {
      "get": {
        "operationId": "data_freshness_api_v1_meta_data_freshness__ticker__get",
        "parameters": [
          {
            "in": "path",
            "name": "ticker",
            "required": true,
            "schema": {
              "title": "Ticker",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataFreshnessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Data Freshness",
        "tags": [
          "Health",
          "Meta"
        ]
      }
    },
...
  }
}